Use guint for array size everywhere
Using a GPtrArray for underlying storage means that our length is
restricted to what a guint can represent, so there's no point in
pretending we can do better.

Although it would kind of be nice to use a long and have -1 as an
error value. Nobody uses 32-bit anymore, right?