ArrayList: Fire notify signals when adding items
I'm not really sure about this one. If we're going to have
properties, they really ought to fire notifications when
they chance so that people can (for example) hook them up
to GBindings.

But on the other hand, if we're going to do that then why
not go the whole hog and have item-added, item-removed etc
signals?