ArrayList: Don't take a copy of elements
The idea of using a copy function was taken from Vala's
handling of generics. However, it turns out that from C
it's easier not to do so. Regular GObjects etc start off
with a reference count of 1, so you can create them
straight into an array with a g_object_unref destroy func.